Better/fixed support for huge keys/values.
Change the WT_ITEM::size field from type uint32_t to size_t.

Change test/format configuration to support large key/value items.

Ref: #808 and #70.

+++ src/btree/bt_handle.c
@@ -99,7 +99,7 @@ __wt_btree_open(WT_SESSION_IMPL *session, const char *op_cfg[])
 		 * checkpoint is for an empty file).
 		 */
 		WT_ERR(bm->checkpoint_load(bm, session,
-		    ckpt.raw.data, ckpt.raw.size,
+		    ckpt.raw.data, (uint32_t)ckpt.raw.size,
 		    root_addr, &root_addr_size, readonly));
 		if (creation || root_addr_size == 0)
 			WT_ERR(__btree_tree_open_empty(session, creation));
